Build fixes
git-svn-id: https://devel.codership.com/svn/galera/trunk@1076 9148e567-b586-4176-b9b8-4e36cb4ec61b

+++ galeracomm/common/src/pollpp.cpp
@@ -145,7 +145,7 @@ int PollDef::poll(const int timeout)
         long long tstamp = PollContext::get_timestamp();
         bool check_timeout = (tstamp - last_check) >= Poll::DEFAULT_KA_INTERVAL;
 
-        wait = timeout - (tstamp - begin);
+        wait = timeout - static_cast<int>(tstamp - begin);
 
         if (p_ret == -1 && err == EINTR) p_ret = 0;
